api/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
httpx==0.26.0

# AI Services
google-generativeai==0.8.3
anthropic==0.18.1  # fallback

# Date handling
python-dateutil==2.8.2

# CORS & Security
python-dotenv==1.0.0

# Logging
structlog==24.1.0

# Testing (dev)
pytest==7.4.4
pytest-asyncio==0.23.3

